Prevent Marker from leaking memory
Keeping a static reference to DefaultInfoWindow prevents it from being GC'ed.
This is especially problematic as this in turn holds a reference to the parent
MapView. This leads to leaking the complete View hierarchy when the fragment
is detached.
To fix this, null out the reference when the parent MapView is detached.
In case of the default icon this is not as troublesome since Drawables only keep
a weak reference, but why not clean it up if possible.